improve(config): allow most template strings in templated module inputs
Basically we now defer the full resolution on input template strings,
with the caveat that input values needed for module identifiers need
to be resolvable immediately.

This involved quite a bit of added complexity, which is prompting me to
seriously a refactor in our template string resolution logic. This does
the job for now, but I think we'll soon want to implement at least
partial lazy+recursive resolution of template strings, instead of all
this fine-grained logic to control what's resolved when.